requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.40.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.3
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
PyJWT==2.6.0
pyparsing==3.1.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
seaborn==0.12.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
